Split out the typescript of the conversion tests into a seperate directory
